+++ src/main/server.ts
@@ -1,7 +1,4 @@
 #!/usr/bin/env node
-const enableAppInsights = require('./app-insights/app-insights');
-enableAppInsights();
-
 const { Logger } = require('@hmcts/nodejs-logging');
 import { app } from './app';
