+++ src/test/app/app-insights/app-insights.test.ts
@@ -1,3 +1,27 @@
+type AppInsightsMock = {
+  setup: jest.Mock;
+  setSendLiveMetrics: jest.Mock;
+  start: jest.Mock;
+  trackTrace: jest.Mock;
+  tags: Record<string, string>;
+};
+
+const createMocks = (): AppInsightsMock => {
+  const start = jest.fn();
+  const setSendLiveMetrics = jest.fn().mockReturnValue({ start });
+  const setup = jest.fn().mockReturnValue({ setSendLiveMetrics });
+  const trackTrace = jest.fn();
+  const tags: Record<string, string> = {};
+
+  return {
+    setup,
+    setSendLiveMetrics,
+    start,
+    trackTrace,
+    tags
+  };
+};
+
 describe('app insights bootstrap', () => {
   afterEach(() => {
     jest.resetModules();
@@ -5,13 +29,20 @@ describe('app insights bootstrap', () => {
   });
 
   it('does nothing when connection string is missing', () => {
-    const setupMock = jest.fn();
+    const mocks = createMocks();
 
     jest.doMock('config', () => ({
       get: jest.fn().mockReturnValue(false)
     }));
     jest.doMock('applicationinsights', () => ({
-      setup: setupMock
+      setup: mocks.setup,
+      defaultClient: {
+        context: {
+          tags: mocks.tags,
+          keys: { cloudRole: 'cloudRole' }
+        },
+        trackTrace: mocks.trackTrace
+      }
     }));
     jest.doMock('@hmcts/nodejs-logging', () => ({
       Logger: {
@@ -24,18 +55,24 @@ describe('app insights bootstrap', () => {
       enableAppInsights();
     });
 
-    expect(setupMock).not.toHaveBeenCalled();
+    expect(mocks.setup).not.toHaveBeenCalled();
   });
 
-  it('starts application insights when connection string exists', () => {
-    const startMock = jest.fn();
-    const setupMock = jest.fn().mockReturnValue({ start: startMock });
+  it('starts application insights and sets cloud role when connection string exists', () => {
+    const mocks = createMocks();
 
     jest.doMock('config', () => ({
       get: jest.fn().mockReturnValue('InstrumentationKey=test-key;IngestionEndpoint=https://test/')
     }));
     jest.doMock('applicationinsights', () => ({
-      setup: setupMock
+      setup: mocks.setup,
+      defaultClient: {
+        context: {
+          tags: mocks.tags,
+          keys: { cloudRole: 'cloudRole' }
+        },
+        trackTrace: mocks.trackTrace
+      }
     }));
     jest.doMock('@hmcts/nodejs-logging', () => ({
       Logger: {
@@ -48,8 +85,11 @@ describe('app insights bootstrap', () => {
       enableAppInsights();
     });
 
-    expect(setupMock).toHaveBeenCalledWith('InstrumentationKey=test-key;IngestionEndpoint=https://test/');
-    expect(startMock).toHaveBeenCalled();
+    expect(mocks.setup).toHaveBeenCalledWith('InstrumentationKey=test-key;IngestionEndpoint=https://test/');
+    expect(mocks.setSendLiveMetrics).toHaveBeenCalledWith(true);
+    expect(mocks.start).toHaveBeenCalled();
+    expect(mocks.tags.cloudRole).toBe('rpe-expressjs-template');
+    expect(mocks.trackTrace).toHaveBeenCalledWith({ message: 'App insights activated' });
   });
 
   it('logs a warning and continues if setup throws', () => {
